订单完善
订单管理，个人信息管理完善，后台添加密码

+++ hll/delete.php
@@ -5,22 +5,22 @@ include_once $mypath.'/includes/db.inc.php';
 include_once $mypath.'/includes/helpers.inc.php';
 session_start();
 
-
-if(isset($_GET['delimg'])){
-    unlink('../'.$_GET['delimg']);
-    $sql = 'DELETE FROM g_image_tbl WHERE url="'.$_GET['delimg'].'"';
+if(isset($_SESSION['login'])) {
+    if (isset($_GET['delimg'])) {
+        unlink('../' . $_GET['delimg']);
+        $sql = 'DELETE FROM g_image_tbl WHERE url="' . $_GET['delimg'] . '"';
 //    $pdo->exec($sql);
-    exeNew($sql);
-    $g_id=$_GET['g_id'];
-    printView('hll/view/goods_edit.html.php');
-    exit;
+        exeNew($sql);
+        $g_id = $_GET['g_id'];
+        printView('hll/view/goods_edit.html.php');
+        exit;
 
+    }
+
+
+    if (isset($_GET['name'])) {
+        echo $_GET['name'];
+        exit;
+    }
 }
-
-
-if(isset($_GET['name'])){
-    echo $_GET['name'];
-    exit;
-}
-
 ?>
